search criteria and ID hashcode/equals

+++ corpdir-srv-jpa/src/main/java/com/github/bordertech/corpdir/jpa/v1/api/OrgUnitServiceImpl.java
@@ -2,22 +2,26 @@ package com.github.bordertech.corpdir.jpa.v1.api;
 
 import com.github.bordertech.corpdir.api.exception.NotFoundException;
 import com.github.bordertech.corpdir.api.exception.ServiceException;
-import com.github.bordertech.corpdir.api.response.ServiceBasicResponse;
-import com.github.bordertech.corpdir.api.response.ServiceResponse;
+import com.github.bordertech.corpdir.api.response.BasicResponse;
+import com.github.bordertech.corpdir.api.response.DataResponse;
 import com.github.bordertech.corpdir.api.v1.OrgUnitService;
 import com.github.bordertech.corpdir.api.v1.model.OrgUnit;
 import com.github.bordertech.corpdir.api.v1.model.Position;
 import com.github.bordertech.corpdir.jpa.common.AbstractJpaService;
-import com.github.bordertech.corpdir.jpa.util.MapperUtil;
 import com.github.bordertech.corpdir.jpa.entity.OrgUnitEntity;
 import com.github.bordertech.corpdir.jpa.entity.PositionEntity;
+import com.github.bordertech.corpdir.jpa.util.EmfUtil;
+import com.github.bordertech.corpdir.jpa.util.MapperUtil;
 import com.github.bordertech.corpdir.jpa.v1.mapper.OrgUnitMapper;
 import com.github.bordertech.corpdir.jpa.v1.mapper.PositionMapper;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import javax.inject.Singleton;
 import javax.persistence.EntityManager;
 import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Predicate;
 import javax.persistence.criteria.Root;
 
 /**
@@ -33,26 +37,42 @@ public class OrgUnitServiceImpl extends AbstractJpaService implements OrgUnitSer
 	private static final PositionMapper POSITION_MAPPER = new PositionMapper();
 
 	@Override
-	public ServiceResponse<List<OrgUnit>> getOrgUnits(final String search, final Boolean topLevel) {
+	public DataResponse<List<OrgUnit>> getOrgUnits(final String search, final Boolean assigned) {
 		EntityManager em = getEntityManager();
 		try {
 			CriteriaBuilder cb = em.getCriteriaBuilder();
 			CriteriaQuery<OrgUnitEntity> qry = cb.createQuery(OrgUnitEntity.class);
 
-			// TODO Implement Search criteria and topOnly
 			Root<OrgUnitEntity> from = qry.from(OrgUnitEntity.class);
 			qry.select(from);
 
+			// Search
+			List<Predicate> preds = new ArrayList<>();
+			if (search != null && !search.isEmpty()) {
+				preds.add(EmfUtil.createSearchTextCriteria(cb, from, search));
+			}
+			// Assigned
+			if (assigned != null) {
+				preds.add(EmfUtil.createAssignedCriteria(cb, from, assigned, "parentOrgUnit"));
+			}
+			// AND
+			if (!preds.isEmpty()) {
+				qry.where(EmfUtil.createAndCriteria(cb, preds));
+			}
+
+			// Order by
+			qry.orderBy(EmfUtil.getDefaultOrderBy(cb, from));
+
 			List<OrgUnitEntity> rows = em.createQuery(qry).getResultList();
 			List<OrgUnit> list = ORGUNIT_MAPPER.convertEntitiesToApis(em, rows);
-			return new ServiceResponse<>(list);
+			return new DataResponse<>(list);
 		} finally {
 			em.close();
 		}
 	}
 
 	@Override
-	public ServiceResponse<OrgUnit> getOrgUnit(final String orgUnitKeyId) {
+	public DataResponse<OrgUnit> getOrgUnit(final String orgUnitKeyId) {
 		EntityManager em = getEntityManager();
 		try {
 			OrgUnitEntity entity = getOrgUnitEntity(em, orgUnitKeyId);
@@ -63,7 +83,7 @@ public class OrgUnitServiceImpl extends AbstractJpaService implements OrgUnitSer
 	}
 
 	@Override
-	public ServiceResponse<OrgUnit> createOrgUnit(final OrgUnit orgUnit) {
+	public DataResponse<OrgUnit> createOrgUnit(final OrgUnit orgUnit) {
 		EntityManager em = getEntityManager();
 		try {
 			MapperUtil.checkApiIDsForCreate(orgUnit);
@@ -83,12 +103,22 @@ public class OrgUnitServiceImpl extends AbstractJpaService implements OrgUnitSer
 	}
 
 	@Override
-	public ServiceResponse<OrgUnit> updateOrgUnit(final String orgUnitKeyId, final OrgUnit orgUnit) {
+	public DataResponse<OrgUnit> updateOrgUnit(final String orgUnitKeyId, final OrgUnit orgUnit) {
 		EntityManager em = getEntityManager();
 		try {
 			em.getTransaction().begin();
 			OrgUnitEntity entity = getOrgUnitEntity(em, orgUnitKeyId);
 			MapperUtil.checkIdentifiersMatch(orgUnit, entity);
+			// Only the ROOT OU can have itself as the parent
+			boolean root = EmfUtil.isRootOrgUnit(entity);
+			if (!root) {
+				if (Objects.equals(orgUnit.getId(), orgUnit.getParentId())) {
+					throw new IllegalArgumentException("Only the Root OU can have itself as a parent OU.");
+				}
+				if (orgUnit.getSubIds().contains(orgUnit.getId())) {
+					throw new IllegalArgumentException("Only the Root OU can have itself as a child OU.");
+				}
+			}
 			ORGUNIT_MAPPER.copyApiToEntity(em, orgUnit, entity);
 			em.getTransaction().commit();
 			return buildResponse(em, entity);
@@ -98,33 +128,47 @@ public class OrgUnitServiceImpl extends AbstractJpaService implements OrgUnitSer
 	}
 
 	@Override
-	public ServiceBasicResponse deleteOrgUnit(final String orgUnitKeyId) {
+	public BasicResponse deleteOrgUnit(final String orgUnitKeyId) {
 		EntityManager em = getEntityManager();
 		try {
 			em.getTransaction().begin();
 			OrgUnitEntity entity = getOrgUnitEntity(em, orgUnitKeyId);
+			// Cannot delete the ROOT OU
+			if (EmfUtil.isRootOrgUnit(entity)) {
+				throw new IllegalArgumentException("Cannot delete the ROOT OU.");
+			}
 			em.remove(entity);
 			em.getTransaction().commit();
-			return new ServiceBasicResponse();
+			return new BasicResponse();
 		} finally {
 			em.close();
 		}
 	}
 
 	@Override
-	public ServiceResponse<List<OrgUnit>> getSubOrgUnits(final String orgUnitKeyId) {
+	public DataResponse<List<OrgUnit>> getSubOrgUnits(final String orgUnitKeyId) {
 		EntityManager em = getEntityManager();
 		try {
 			OrgUnitEntity entity = getOrgUnitEntity(em, orgUnitKeyId);
 			List<OrgUnit> list = ORGUNIT_MAPPER.convertEntitiesToApis(em, entity.getSubOrgUnits());
-			return new ServiceResponse<>(list);
+			// For the ROOT OU - Dont send itself as a sub OU
+			if (EmfUtil.isRootOrgUnit(entity)) {
+				String rootId = MapperUtil.convertEntityIdforApi(entity);
+				for (OrgUnit ou : list) {
+					if (Objects.equals(ou.getId(), rootId)) {
+						list.remove(ou);
+						break;
+					}
+				}
+			}
+			return new DataResponse<>(list);
 		} finally {
 			em.close();
 		}
 	}
 
 	@Override
-	public ServiceResponse<OrgUnit> addSubOrgUnit(final String orgUnitKeyId, final String subOrgUnitKeyId) {
+	public DataResponse<OrgUnit> addSubOrgUnit(final String orgUnitKeyId, final String subOrgUnitKeyId) {
 		EntityManager em = getEntityManager();
 		try {
 			em.getTransaction().begin();
@@ -132,6 +176,10 @@ public class OrgUnitServiceImpl extends AbstractJpaService implements OrgUnitSer
 			OrgUnitEntity entity = getOrgUnitEntity(em, orgUnitKeyId);
 			// Get the sub org unit
 			OrgUnitEntity subOrgUnit = getOrgUnitEntity(em, subOrgUnitKeyId);
+			// Cant add an OU to itself
+			if (Objects.equals(entity, subOrgUnit)) {
+				throw new IllegalArgumentException("Cannot add an OU to itself");
+			}
 			// Remove Sub Org Unit from its OLD parent (if it had one)
 			OrgUnitEntity oldParent = subOrgUnit.getParentOrgUnit();
 			if (oldParent != null) {
@@ -147,14 +195,21 @@ public class OrgUnitServiceImpl extends AbstractJpaService implements OrgUnitSer
 	}
 
 	@Override
-	public ServiceResponse<OrgUnit> removeSubOrgUnit(final String orgUnitKeyId, final String subOrgUnitKeyId) {
+	public DataResponse<OrgUnit> removeSubOrgUnit(final String orgUnitKeyId, final String subOrgUnitKeyId) {
 		EntityManager em = getEntityManager();
 		try {
+			if (Objects.equals(orgUnitKeyId, subOrgUnitKeyId)) {
+				throw new IllegalArgumentException("Cannot remove an OU from itself");
+			}
 			em.getTransaction().begin();
 			// Get the org unit
 			OrgUnitEntity entity = getOrgUnitEntity(em, orgUnitKeyId);
 			// Get the sub org unit
 			OrgUnitEntity subOrgUnit = getOrgUnitEntity(em, subOrgUnitKeyId);
+			// Cant add an OU to itself
+			if (Objects.equals(entity, subOrgUnit)) {
+				throw new IllegalArgumentException("Cannot add an OU to itself");
+			}
 			// Remove the sub org unit
 			entity.removeSubOrgUnit(subOrgUnit);
 			em.getTransaction().commit();
@@ -165,19 +220,19 @@ public class OrgUnitServiceImpl extends AbstractJpaService implements OrgUnitSer
 	}
 
 	@Override
-	public ServiceResponse<List<Position>> getPositions(final String orgUnitKeyId) {
+	public DataResponse<List<Position>> getPositions(final String orgUnitKeyId) {
 		EntityManager em = getEntityManager();
 		try {
 			OrgUnitEntity entity = getOrgUnitEntity(em, orgUnitKeyId);
 			List<Position> list = POSITION_MAPPER.convertEntitiesToApis(em, entity.getPositions());
-			return new ServiceResponse<>(list);
+			return new DataResponse<>(list);
 		} finally {
 			em.close();
 		}
 	}
 
 	@Override
-	public ServiceResponse<OrgUnit> addPosition(final String orgUnitKeyId, final String positionKeyId) {
+	public DataResponse<OrgUnit> addPosition(final String orgUnitKeyId, final String positionKeyId) {
 		EntityManager em = getEntityManager();
 		try {
 			em.getTransaction().begin();
@@ -199,7 +254,7 @@ public class OrgUnitServiceImpl extends AbstractJpaService implements OrgUnitSer
 	}
 
 	@Override
-	public ServiceResponse<OrgUnit> removePosition(final String orgUnitKeyId, final String positionKeyId) {
+	public DataResponse<OrgUnit> removePosition(final String orgUnitKeyId, final String positionKeyId) {
 		EntityManager em = getEntityManager();
 		try {
 			em.getTransaction().begin();
@@ -217,12 +272,12 @@ public class OrgUnitServiceImpl extends AbstractJpaService implements OrgUnitSer
 	}
 
 	@Override
-	public ServiceResponse<Position> getManagerPosition(final String orgUnitKeyId) {
+	public DataResponse<Position> getManagerPosition(final String orgUnitKeyId) {
 		EntityManager em = getEntityManager();
 		try {
 			OrgUnitEntity entity = getOrgUnitEntity(em, orgUnitKeyId);
 			Position data = POSITION_MAPPER.convertEntityToApi(em, entity.getManagerPosition());
-			return new ServiceResponse<>(data);
+			return new DataResponse<>(data);
 		} finally {
 			em.close();
 		}
@@ -234,9 +289,9 @@ public class OrgUnitServiceImpl extends AbstractJpaService implements OrgUnitSer
 	 * @param entity the org unit entity
 	 * @return the service response with org unit API object
 	 */
-	protected ServiceResponse<OrgUnit> buildResponse(final EntityManager em, final OrgUnitEntity entity) {
+	protected DataResponse<OrgUnit> buildResponse(final EntityManager em, final OrgUnitEntity entity) {
 		OrgUnit data = ORGUNIT_MAPPER.convertEntityToApi(em, entity);
-		return new ServiceResponse<>(data);
+		return new DataResponse<>(data);
 	}
 
 	/**
